requirement.txt
aiohttp==3.8.5
aiosignal==1.3.1
altgraph==0.17.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
auto-py-to-exe==2.39.0
backcall==0.2.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
dateparser==1.1.8
decorator==5.1.1
dill==0.3.7
Eel==0.16.0
exceptiongroup==1.1.3
executing==1.2.0
fonttools==4.42.1
frozenlist==1.4.0
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==2.0.2
idna==3.4
imageio==2.31.3
ipython==8.15.0
ipywidgets==8.1.1
jedi==0.19.0
joblib==1.3.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
llvmlite==0.39.1
matplotlib==3.7.3
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
numba==0.56.4
numpy==1.23.5
p5py==1.0.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pefile==2023.2.7
pep517==0.13.0
pickleshare==0.7.5
Pillow==10.0.0
plotly==5.16.1
prompt-toolkit==3.0.39
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.18.0
Pygments==2.16.1
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
pyparsing==3.1.1
python-binance==1.0.19
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
regex==2023.8.8
requests==2.31.0
schedule==1.2.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
stack-data==0.6.2
TA-Lib
tenacity==8.2.3
threadpoolctl==3.2.0
tomli==2.0.1
tqdm==4.66.1
traitlets==5.10.0
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==2.0.4
vectorbt==0.25.5
wcwidth==0.2.6
websockets==11.0.3
whichcraft==0.6.1
widgetsnbextension==4.0.9
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
